requirements.txt
asgiref
beautifulsoup4
bleach
certifi
charset-normalizer
cloudinary
Django
django-bootstrap5
django-ckeditor
django-cloudinary-storage
django-jazzmin
django-js-asset
django-summernote
django-user-agents
Faker
gunicorn
idna
itsdangerous
paystack
Pillow
protobuf
psycopg2-binary
python-dateutil
python-decouple
requests
shortuuid
simplejson
six
soupsieve
sqlparse
tzdata
ua-parser
urllib3
user-agents
webencodings

whitenoise

